backend/requirements.txt
alembic>=1.15.1             # 数据库迁移
APScheduler>=3.11.0         # 定时任务
fastapi>=0.115.2            # fastapi 框架
Jinja2>=3.1.6               # 模板引擎
typer>=0.9.0                # 命令行工具
click>=8.1.7                # 命令行参数解析
uvicorn>=0.30.6             # uvicorn web 框架
gunicorn>=23.0.0            # 协程框架
websockets>=14.2            # websocket 框架
httpx>=0.28.1               # HTTP 客户端
croniter>=6.0.0             # 实现cron表达式验证和解析执行计划
pandas>=2.2.2               # 数据处理
openpyxl>=3.1.5             # Excel
SQLAlchemy>=2.0.36          # 数据库ORM
pillow>=11.0.0              # 图片处理
passlib>=1.7.4              # 密码加密
user-agents>=2.2.0          # 获取用户UA
PyJWT>=2.9.0                # OAuth2
pydantic-settings>=2.5.2    # 配置设置
psutil>=6.1.0               # 系统信息
python-multipart>=0.0.9     # request.form() 对表单进行「解析」时安装
greenlet>=3.1.1             # 协程框架
bcrypt>=4.0.1               # 密码加密解析,切勿升级，如果升级，请同时升级python版本
itsdangerous>=2.2.0         # 用于安全处理各种数据，如密码、密钥等
aiofiles>=24.1.0            # 文件操作
redis>=5.2.1                # redis 同/异步操作数据库(用户celery配套使用)redis 异步操作数据库 redis已经完全具备了aioredis的功能，无需重复安全,且aioredis已经不再维护也不兼容3.10+的版本
asyncmy>=0.2.9              # mysql 异步操作数据库：基于 mysqlclient：asyncmy 是 mysqlclient 的异步版本，mysqlclient 是一个 C 语言编写的 MySQL 客户端，性能较高。性能：asyncmy 通常在性能上优于 aiomysql，特别是在高并发和大数据量的场景下。
PyMySQL>=1.1.2              # mysql 同步步操作数据库基于 pymysql：aiomysql 是 pymysql 的异步版本，pymysql 是一个纯 Python 实现的 MySQL 客户端。成熟度：aiomysql 相对较为成熟，社区支持较好，文档也比较完善。
asyncpg>=0.30.0             # postgresql 异步操作数据库基于 psycopg2：asyncpg 是 psycopg2 的异步版本，psycopg2 是一个 pure-Python PostgreSQL 数据库适配器。性能：asyncpg 通常在性能上优于 psycopg2，特别是在高并发和大数据量的场景下。
psycopg2>=2.9.10            # postgresql 同步操作数据库基于 psycopg2：psycopg2 是一个 pure-Python PostgreSQL 适配器。
cryptography>=45.0.2        # mysql8 密码加密
openai>=1.55.2              # ai 大模型
rich>=13.9.4                # 终端打印美化
sqlglot[rs]>=27.8.0         # sql 解析
pydantic_validation_decorator>=0.1.4    # 模型验证
loguru>=0.7.3
fastapi-limiter>=0.1.6

# amqp>=5.3.1
# python-socketio>=5.14.3

langchain>=1.1.0                   # langchain 框架
langchain_openai            # langchain openai 集成
langchain_community         # langchain community extensions
json_repair                 # JSON 修复

alibabacloud-dingtalk>=2.0.0    # 阿里云钉钉 SDK
alibabacloud-tea-openapi>=0.3.0 # 阿里云 OpenAPI SDK
alibabacloud-tea-util>=0.3.0    # 阿里云 OpenAPI 工具包

langfuse>=0.6.5                # Langfuse SDK 